requirements.txt
asgiref==3.5.2
Django==4.1.1
django-environ==0.9.0
djangorestframework==3.13.1
drf-yasg==1.21.3
pytz==2022.2.1
sqlparse==0.4.2
psycopg2-binary==2.9.3
PyYAML==6.0
gunicorn==20.1.0
django-cors-headers==3.13.0